Refactoring arhitecture look for project
