+++ handlers/nextdate.go
@@ -1,0 +1,28 @@
+package handlers
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/akonovalovdev/go_final_project/utils"
+)
+
+func NextDateHandler(w http.ResponseWriter, r *http.Request) {
+	nowStr := r.FormValue("now")
+	dateStr := r.FormValue("date")
+	repeatStr := r.FormValue("repeat")
+
+	now, err := time.Parse("20060102", nowStr)
+	if err != nil {
+		http.Error(w, "Invalid now parameter", http.StatusBadRequest)
+		return
+	}
+
+	nextDate, err := utils.NextDate(now, dateStr, repeatStr)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	w.Write([]byte(nextDate))
+}
