+++ models/tasks.go
@@ -1,0 +1,8 @@
+package models
+
+type Task struct {
+	Date    string `json:"date"`
+	Title   string `json:"title"`
+	Comment string `json:"comment"`
+	Repeat  string `json:"repeat"`
+}
